# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AddressEnrichmentOut": {
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "canonical_address": {
            "title": "Canonical Address",
            "type": "string"
          },
          "data_mode": {
            "title": "Data Mode",
            "type": "string"
          },
          "external_address_id": {
            "format": "uuid",
            "title": "External Address Id",
            "type": "string"
          },
          "municipality_code": {
            "title": "Municipality Code",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "postal_town": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Town"
          },
          "register_unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegisterUnitReferenceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_easting": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Source Easting",
            "type": "string"
          },
          "source_northing": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Source Northing",
            "type": "string"
          },
          "source_srid": {
            "title": "Source Srid",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "source",
          "data_mode",
          "external_address_id",
          "canonical_address",
          "municipality_code",
          "postal_code",
          "postal_town",
          "status",
          "source_srid",
          "source_easting",
          "source_northing",
          "attribution",
          "observed_at",
          "updated_at"
        ],
        "title": "AddressEnrichmentOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LifecycleReadinessOut": {
        "properties": {
          "considered_run_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Considered Run Ids",
            "type": "array"
          },
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "latest_scheduled_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Scheduled Run At"
          },
          "qualifying_runs": {
            "title": "Qualifying Runs",
            "type": "integer"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "required_runs": {
            "title": "Required Runs",
            "type": "integer"
          },
          "rule_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Version"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "source_key",
          "status",
          "required_runs",
          "qualifying_runs",
          "rule_version",
          "reasons",
          "considered_run_ids",
          "latest_scheduled_run_at",
          "evaluated_at"
        ],
        "title": "LifecycleReadinessOut",
        "type": "object"
      },
...
      "PropertyOut": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "address_enrichments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AddressEnrichmentOut"
            },
            "title": "Address Enrichments",
            "type": "array"
          },
          "area_m2": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area M2"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "city_id": {
            "title": "City Id",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "municipality_code": {
            "title": "Municipality Code",
            "type": "string"
          },
          "new_construction": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Construction"
          },
          "property_type": {
            "title": "Property Type",
            "type": "string"
          },
          "rooms": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rooms"
          },
          "unit_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Identifier"
          }
        },
        "required": [
          "id",
          "address",
          "city_id",
          "city",
          "municipality_code",
          "property_type",
          "unit_identifier",
          "rooms",
          "area_m2",
          "new_construction",
          "latitude",
          "longitude"
        ],
        "title": "PropertyOut",
        "type": "object"
      },
...
      "RentalListingClassificationOut": {
        "properties": {
          "confidence": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "listing_id": {
            "format": "uuid",
            "title": "Listing Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "rule_version": {
            "title": "Rule Version",
            "type": "string"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "listing_id",
          "tag",
          "confidence",
          "reason",
          "evidence",
          "rule_version",
          "created_at",
          "updated_at"
        ],
        "title": "RentalListingClassificationOut",
        "type": "object"
      },
      "RentalListingList": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RentalListingOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "RentalListingList",
        "type": "object"
      },
...
      "ScoreRunOut": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "city_id": {
            "title": "City Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "definition_set_hash": {
            "title": "Definition Set Hash",
            "type": "string"
          },
          "feature_schema_revision": {
            "title": "Feature Schema Revision",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "population_fingerprint": {
            "title": "Population Fingerprint",
            "type": "string"
          },
          "population_size": {
            "title": "Population Size",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "city_id",
          "as_of_date",
          "feature_schema_revision",
          "definition_set_hash",
          "population_fingerprint",
          "population_size",
          "status",
          "created_at"
        ],
        "title": "ScoreRunOut",
        "type": "object"
      },
...
      "SignalDimensionSetOut": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/SignalDimensionEvaluationOut"
            },
            "title": "Dimensions",
            "type": "array"
          },
          "feature_snapshot_hash": {
            "title": "Feature Snapshot Hash",
            "type": "string"
          },
          "feature_snapshot_id": {
            "format": "uuid",
            "title": "Feature Snapshot Id",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "signal_id": {
            "format": "uuid",
            "title": "Signal Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "signal_id",
          "as_of_date",
          "feature_snapshot_id",
          "feature_snapshot_hash",
          "dimensions"
        ],
        "title": "SignalDimensionSetOut",
        "type": "object"
      },
...
      "SignalOutcomeOut": {
        "properties": {
          "confidence": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dedupe_key": {
            "title": "Dedupe Key",
            "type": "string"
          },
          "event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "listing_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Id"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "outcome_type": {
            "title": "Outcome Type",
            "type": "string"
          },
          "rule_version": {
            "title": "Rule Version",
            "type": "string"
          },
          "signal_id": {
            "format": "uuid",
            "title": "Signal Id",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "verification_level": {
            "title": "Verification Level",
            "type": "string"
          }
        },
        "required": [
          "id",
          "signal_id",
          "listing_id",
          "event_id",
          "outcome_type",
          "subject",
          "verification_level",
          "observed_at",
          "confidence",
          "evidence",
          "rule_version",
          "dedupe_key",
          "created_at",
          "updated_at"
        ],
        "title": "SignalOutcomeOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FlyttSignal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/properties/{property_id}": {
      "get": {
        "operationId": "getProperty",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Property"
      }
    },
...
    "/api/rental-listings": {
      "get": {
        "operationId": "listRentalListings",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "ACTIVE",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "data_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "fixture",
                    "live"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Mode"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalListingList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rental Listings"
      }
    },
    "/api/rental-listings/{listing_id}/classifications": {
      "get": {
        "operationId": "listRentalListingClassifications",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RentalListingClassificationOut"
                  },
                  "title": "Response Listrentallistingclassifications",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rental Listing Classifications"
      }
    },
...
    "/api/score-runs": {
      "get": {
        "operationId": "listScoreRuns",
        "parameters": [
          {
            "in": "query",
            "name": "city_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScoreRunOut"
                  },
                  "title": "Response Listscoreruns",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Score Runs"
      }
    },
    "/api/score-runs/{run_id}": {
      "get": {
        "operationId": "getScoreRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Score Run"
      }
    },
    "/api/score-runs/{run_id}/signals/{signal_id}": {
      "get": {
        "operationId": "getScoreRunSignalDimensions",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "signal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Signal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalDimensionSetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Score Run Signal Dimensions"
      }
    },
...
    "/api/signals/{signal_id}/outcomes": {
      "get": {
        "operationId": "listSignalOutcomes",
        "parameters": [
          {
            "in": "path",
            "name": "signal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Signal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SignalOutcomeOut"
                  },
                  "title": "Response Listsignaloutcomes",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Signal Outcomes"
      }
    },
    "/api/source-runs": {
      "get": {
        "operationId": "listSourceRuns",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "RUNNING",
                    "SUCCESS",
                    "FAILED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceRunOut"
                  },
                  "title": "Response Listsourceruns",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Source Runs"
      }
    },
...
    "/api/sources/{source_key}/lifecycle-readiness": {
      "get": {
        "operationId": "getSourceLifecycleReadiness",
        "parameters": [
          {
            "in": "path",
            "name": "source_key",
            "required": true,
            "schema": {
              "title": "Source Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleReadinessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source Lifecycle Readiness"
      }
    },
...
  }
}